Collect visible node indices inside the selection rectangle.

+++ shaders/mouse_picking.frag
@@ -1,12 +1,11 @@
 #version 450
-#extension GL_EXT_shader_16bit_storage : require
 
 layout (set = 0, binding = 0, input_attachment_index = 0) uniform usubpassInput inputNodeIndex;
 layout (set = 0, binding = 1, std430) writeonly buffer ResultBuffer {
-    uint16_t nodeIndex;
+    uint nodeIndex;
 };
 
 void main() {
     // This shader is expected to be executed only once (scissor=1x1), therefore synchronization (such as atomic operations) is not needed.
-    nodeIndex = uint16_t(subpassLoad(inputNodeIndex).r);
+    nodeIndex = subpassLoad(inputNodeIndex).r;
 }
